move icons to lib
