--- dotcom-rendering/src/lib/worldCup2026.ts
@@ -1,7 +1,0 @@
-export const worldCup2026PageIds = [
-	'football/world-cup-2026',
-	'football/world-cup-2026/fixtures',
-	'football/world-cup-2026/overview',
-];
-
-export const worldCupTagIds = [];
